requirements.txt
absl-py==1.4.0
array-record==0.5.0
asttokens==2.4.1
astunparse==1.6.3
cachetools==5.3.2
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
contourpy==1.2.0
cycler==0.12.1
decorator==5.1.1
dm-tree==0.1.8
etils==1.6.0
exceptiongroup==1.2.0
executing==2.0.1
flatbuffers==23.5.26
fonttools==4.46.0
fsspec==2023.12.1
gast==0.5.4
google-auth==2.25.2
google-auth-oauthlib==1.1.0
google-pasta==0.2.0
googleapis-common-protos==1.62.0
grpcio==1.60.0
h5py==3.10.0
idna==3.6
importlib-resources==6.1.1
ipython==8.18.1
jedi==0.19.1
keras==2.15.0
kiwisolver==1.4.5
libclang==16.0.6
Markdown==3.5.1
MarkupSafe==2.1.3
matplotlib==3.8.2
matplotlib-inline==0.1.6
ml-dtypes==0.2.0
numpy==1.26.2
nvidia-cublas-cu12==12.2.5.6
nvidia-cuda-cupti-cu12==12.2.142
nvidia-cuda-nvcc-cu12==12.2.140
nvidia-cuda-nvrtc-cu12==12.2.140
nvidia-cuda-runtime-cu12==12.2.140
nvidia-cudnn-cu12==8.9.4.25
nvidia-cufft-cu12==11.0.8.103
nvidia-curand-cu12==10.3.3.141
nvidia-cusolver-cu12==11.5.2.141
nvidia-cusparse-cu12==12.1.2.141
nvidia-nccl-cu12==2.16.5
nvidia-nvjitlink-cu12==12.2.140
oauthlib==3.2.2
opencv-python==4.8.1.78
opt-einsum==3.3.0
packaging==23.2
pandas==2.1.4
parso==0.8.3
pexpect==4.9.0
Pillow==10.1.0
promise==2.3
prompt-toolkit==3.0.41
protobuf==3.20.3
psutil==5.9.6
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.5.1
pyasn1-modules==0.3.0
Pygments==2.17.2
pyparsing==3.1.1
python-dateutil==2.8.2
pytz==2023.3.post1
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
scipy==1.11.4
six==1.16.0
stack-data==0.6.3
tensorboard==2.15.1
tensorboard-data-server==0.7.2
tensorflow==2.15.0.post1
tensorflow-datasets==4.9.3
tensorflow-estimator==2.15.0
tensorflow-examples @ git+https://github.com/tensorflow/examples.git@0b14ce1c88537b94772cc99d995d936417be6f5d
tensorflow-io-gcs-filesystem==0.34.0
tensorflow-metadata==1.14.0
termcolor==2.4.0
toml==0.10.2
tqdm==4.66.1
traitlets==5.14.0
typing_extensions==4.9.0
tzdata==2023.3
urllib3==2.1.0
wcwidth==0.2.12
Werkzeug==3.0.1
wrapt==1.14.1
zipp==3.17.0
